# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lyra personal productivity API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: date
          in: query
          schema:
            type: ["string", "null"]
          description: Filter by due date (YYYY-MM-DD)
        - name: status
          in: query
          schema:
            type: ["string", "null"]
          description: Filter by status (pending, done, all)
        - name: priority
          in: query
          schema:
            type: ["string", "null"]
          description: Filter by priority (none, low, medium, high)
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /prayers:
    get:
      operationId: listPrayers
      tags: [prayers]
      summary: List prayer logs for a date
      parameters:
        - name: date
          in: query
          schema:
            type: ["string", "null"]
          description: Date filter (YYYY-MM-DD, defaults to today)
      responses:
        "200":
          description: List of prayer entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prayer"
  /prayers/{id}:
    patch:
      operationId: updatePrayer
      tags: [prayers]
      summary: Update a prayer's status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePrayerBody"
      responses:
        "200":
          description: Updated prayer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prayer"
  /prayers/seed:
    post:
      operationId: seedPrayers
      tags: [prayers]
      summary: Seed today's 5 prayers if not yet seeded
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeedPrayersBody"
      responses:
        "200":
          description: Seeded or existing prayers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prayer"

  # ─── FOCUS SESSIONS ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, done]
        priority:
          type: string
          enum: [none, low, medium, high]
        dueDate:
          type: ["string", "null"]
        dueTime:
          type: ["string", "null"]
        list:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - status
        - priority
        - tags
        - createdAt
        - updatedAt
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
          enum: [none, low, medium, high]
        dueDate:
          type: ["string", "null"]
        dueTime:
          type: ["string", "null"]
        list:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
      required:
        - title
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        dueTime:
          type: ["string", "null"]
        list:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
...
    Prayer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          enum: [Fajr, Dhuhr, Asr, Maghrib, Isha]
        date:
          type: string
        scheduledTime:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, on_time, late, missed]
        completedAt:
          type: ["string", "null"]
      required:
        - id
        - name
        - date
        - status
    UpdatePrayerBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, on_time, late, missed]
        scheduledTime:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
      required:
        - status
    SeedPrayersBody:
      type: object
      properties:
        date:
          type: string
          description: Date in YYYY-MM-DD format
        times:
          type: object
          description: Prayer times keyed by prayer name
          properties:
            Fajr:
              type: ["string", "null"]
            Dhuhr:
              type: ["string", "null"]
            Asr:
              type: ["string", "null"]
            Maghrib:
              type: ["string", "null"]
            Isha:
              type: ["string", "null"]
      required:
        - date

    # ─── FOCUS SESSION SCHEMAS ──────────────────────────────────────────────────
...